+++ src/control_plane/migration/postgres/mod.rs
@@ -3,8 +3,11 @@ mod tests;
 
 mod backup_postgres_database;
 mod postgres_backup_options;
+mod postgres_migration_operations;
+mod postgres_migration_operations_options;
 mod postgres_provision_target_options;
 mod postgres_restore_options;
+mod postgres_source_retirement;
 mod postgres_verify_target_options;
 mod provision_postgres_target;
 mod restore_postgres_database;
@@ -12,8 +15,11 @@ mod verify_postgres_target;
 
 pub(crate) use backup_postgres_database::backup_postgres_database;
 pub(crate) use postgres_backup_options::PostgresBackupOptions;
+pub(crate) use postgres_migration_operations::PostgresMigrationOperations;
+pub(crate) use postgres_migration_operations_options::PostgresMigrationOperationsOptions;
 pub(crate) use postgres_provision_target_options::PostgresProvisionTargetOptions;
 pub(crate) use postgres_restore_options::PostgresRestoreOptions;
+pub(crate) use postgres_source_retirement::PostgresSourceRetirement;
 pub(crate) use postgres_verify_target_options::PostgresVerifyTargetOptions;
 pub(crate) use provision_postgres_target::provision_postgres_target;
 pub(crate) use restore_postgres_database::restore_postgres_database;